+++ bin/cli.js
@@ -420,9 +420,8 @@ async function run(oaFile, options) {
 
     const cliOut = [];
     cliLog.unusedActions.forEach(action => {
-      const description = action.description || 'No description provided';
       cliOut.push(
-        `- Target: ${action.target}\n  Type: ${action.update ? 'update' : action.remove ? 'remove' : 'unknown'}`
+        `- Target: ${action.target}\n  Type: ${action.update ? 'update' : action.remove ? 'remove' : action.copy ? 'copy' : 'unknown'}`
       );
     });